Make shutdown more robust by  checking if repository is still around.

+++ trunk/uk.ac.osswatch.simal.web/src/main/java/uk/ac/osswatch/simal/wicket/UserApplication.java
@@ -133,8 +133,10 @@ public class UserApplication extends WebApplication {
 	 * @throws SimalRepositoryException
 	 */
 	public static void destroyRepository() throws SimalRepositoryException {
-		repository.destroy();
-		repository = null;
+	  if (repository != null) {
+	    repository.destroy();
+	    repository = null;
+	  }
 	}
 
 	/**
